+++ dif-data-classes/src/commonMain/kotlin/at/asitplus/dif/InputDescriptor.kt
@@ -1,13 +1,8 @@
-@file:UseSerializers(InputDescriptorSerializer::class)
-
 package at.asitplus.dif
 
-import at.asitplus.dif.rqes.serializers.Base64URLTransactionDataSerializer
-import at.asitplus.dif.rqes.collection_entries.TransactionData
 import com.benasher44.uuid.uuid4
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.UseSerializers
 
 @Serializable(with = InputDescriptorSerializer::class)
 sealed interface InputDescriptor {
@@ -44,22 +39,3 @@ data class DifInputDescriptor(
         constraints = constraints,
     )
 }
-
-@Serializable
-data class QesInputDescriptor(
-    @SerialName("id")
-    override val id: String,
-    @SerialName("group")
-    override val group: String? = null,
-    @SerialName("name")
-    override val name: String? = null,
-    @SerialName("purpose")
-    override val purpose: String? = null,
-    @SerialName("format")
-    override val format: FormatHolder? = null,
-    @SerialName("constraints")
-    override val constraints: Constraint? = null,
-    @SerialName("transaction_data")
-    val transactionData: List<@Serializable(Base64URLTransactionDataSerializer::class) TransactionData>,
-) : InputDescriptor
-
