+++ rqes-data-classes/src/commonMain/kotlin/at/asitplus/rqes/collection_entries/DocumentLocation.kt
@@ -1,6 +1,6 @@
 package at.asitplus.rqes.collection_entries
 
-import at.asitplus.dif.rqes.Method
+import at.asitplus.rqes.Method
 import io.ktor.http.*
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
