+++ rqes-data-classes/src/commonMain/kotlin/at/asitplus/rqes/collection_entries/RqesDocumentDigest.kt
@@ -1,8 +1,8 @@
-package at.asitplus.rqes.collection_entries.DocumentDigestEntries
+package at.asitplus.rqes.collection_entries
 
 import at.asitplus.KmmResult
 import at.asitplus.KmmResult.Companion.wrap
-import at.asitplus.dif.rqes.Method
+import at.asitplus.rqes.Method
 import at.asitplus.signum.indispensable.asn1.ObjectIdSerializer
 import at.asitplus.signum.indispensable.asn1.ObjectIdentifier
 import at.asitplus.signum.indispensable.io.ByteArrayBase64Serializer
