+++ rqes-data-classes/src/commonMain/kotlin/at/asitplus/rqes/serializers/Base64URLTransactionDataSerializer.kt
@@ -1,7 +1,7 @@
 package at.asitplus.rqes.serializers
 
 import at.asitplus.dif.jsonSerializer
-import at.asitplus.dif.rqes.collection_entries.TransactionData
+import at.asitplus.rqes.collection_entries.TransactionData
 import at.asitplus.signum.indispensable.io.Base64UrlStrict
 import io.matthewnelson.encoding.core.Decoder.Companion.decodeToByteArray
 import io.matthewnelson.encoding.core.Encoder.Companion.encodeToString
