+++ rqes-data-classes/src/commonMain/kotlin/at/asitplus/rqes/serializers/CSCSignatureRequestParameterSerializer.kt
@@ -1,8 +1,8 @@
 package at.asitplus.rqes.serializers
 
-import at.asitplus.dif.rqes.CSCSignatureRequestParameters
-import at.asitplus.dif.rqes.SignDocParameters
-import at.asitplus.dif.rqes.SignHashParameters
+import at.asitplus.rqes.CSCSignatureRequestParameters
+import at.asitplus.rqes.SignDocParameters
+import at.asitplus.rqes.SignHashParameters
 import kotlinx.serialization.json.JsonContentPolymorphicSerializer
 import kotlinx.serialization.json.JsonElement
 import kotlinx.serialization.json.jsonObject
